Refactor the execution of commands
The execution of commands is refactored to the individual command types.
This is to make use of polymorphism to eliminate the need for command
type check, improving the readability of code.
