+++ src/main/java/abyss/command/DoneCommand.java
@@ -1,5 +1,7 @@
 package abyss.command;
 
+import java.io.IOException;
+
 import abyss.Abyss;
 import abyss.exception.InvalidCommandException;
 
@@ -31,7 +33,15 @@ public class DoneCommand implements Command {
         this.index = i;
     }
 
-    public int getIndex() {
-        return this.index;
+    /**
+     * Executes the done command.
+     *
+     * @return Response from executing the done command.
+     * @throws IOException If there is error saving to file.
+     */
+    public String execute() throws IOException {
+        String response = Abyss.getTaskManager().markAsDone(index);
+        Abyss.getStorage().saveTasks(Abyss.getTaskManager());
+        return response;
     }
 }
