+++ src/main/java/abyss/command/EventCommand.java
@@ -1,6 +1,11 @@
 package abyss.command;
 
+import java.io.IOException;
+
+import abyss.Abyss;
+import abyss.Ui;
 import abyss.exception.InvalidEventException;
+import abyss.task.Task;
 
 /**
  * Represents a command to add an event to the list of tasks.
@@ -9,7 +14,7 @@ public class EventCommand implements Command {
     private static final String EVENT_REGEX = "^\\S[ -~]*\\/at[ ]+\\S[ -~]*$";
 
     private String description;
-    private String at;
+    private String date;
 
     protected EventCommand(String content) throws InvalidEventException {
         if (!content.matches(EVENT_REGEX)) {
@@ -17,14 +22,24 @@ public class EventCommand implements Command {
         }
         String[] parts = content.split("\\/at[ ]+", 2);
         this.description = parts[0];
-        this.at = parts[1];
+        this.date = parts[1];
     }
 
     public String getDescription() {
         return this.description;
     }
 
-    public String getDate() {
-        return this.at;
+    /**
+     * Executes the add event command.
+     *
+     * @return Response from executing the add event command.
+     * @throws IOException If there is error saving to file.
+     */
+    @Override
+    public String execute() throws IOException {
+        Task task = Abyss.getTaskManager().addEvent(description, date);
+        String response = Ui.replyTaskAdded(task);
+        Abyss.getStorage().saveTasks(Abyss.getTaskManager());
+        return response;
     }
 }
