+++ src/main/java/abyss/command/TodoCommand.java
@@ -1,6 +1,11 @@
 package abyss.command;
 
+import java.io.IOException;
+
+import abyss.Abyss;
+import abyss.Ui;
 import abyss.exception.InvalidTodoException;
+import abyss.task.Task;
 
 /**
  * Represents a command to add a to-do task to the list of tasks.
@@ -17,5 +22,19 @@ public class TodoCommand implements Command {
 
     public String getDescription() {
         return this.description;
+    }
+
+    /**
+     * Executes the add to-do command.
+     *
+     * @return Response from executing the add to-do command.
+     * @throws IOException If there is error saving to file.
+     */
+    @Override
+    public String execute() throws IOException {
+        Task task = Abyss.getTaskManager().addToDo(description);
+        String response = Ui.replyTaskAdded(task);
+        Abyss.getStorage().saveTasks(Abyss.getTaskManager());
+        return response;
     }
 }
